+++ pupil_recording_interface/process/circle_detector.py
@@ -21,6 +21,7 @@ class CircleDetector(BaseProcess):
         circularity=0.8,
         convexity=0.7,
         inertia=0.4,
+        display=True,
         **kwargs,
     ):
         """ Constructor. """
@@ -39,6 +40,7 @@ class CircleDetector(BaseProcess):
             **kwargs,
         )
         self.scale = scale
+        self.display = display
 
     def detect_circle(self, packet):
         """ Detect circle markers. """
@@ -55,6 +57,39 @@ class CircleDetector(BaseProcess):
 
         return circle_markers
 
+    def display_hook(self, packet):
+        """ Add gaze overlay onto frame. """
+        circle_markers = packet["circle_markers"]
+        if circle_markers is None:
+            # Return the attribute to avoid unnecessary waiting
+            return packet.display_frame
+
+        if len(circle_markers) == 0:
+            # Return the attribute to avoid unnecessary waiting
+            return packet.display_frame
+        else:
+            # TODO get the largest marker
+            marker_position = (
+                int(circle_markers[0]["img_pos"][0]),
+                int(circle_markers[0]["img_pos"][1]),
+            )
+
+        frame = packet["display_frame"]
+        if frame.ndim == 2:
+            frame = cv2.cvtColor(frame, cv2.COLOR_GRAY2BGR)
+
+        # TODO make constructor arguments
+        # TODO Define color, radius and thickness in config
+        color = (0, 0, 255)
+        marker_thickness = 5
+        radius = 20
+
+        cv2.circle(
+            frame, marker_position, radius, color, thickness=marker_thickness
+        )
+
+        return frame
+
     def _process_packet(self, packet, block=None):
         """ Process a new packet. """
         packet.circle_markers = self.call(
@@ -62,5 +97,8 @@ class CircleDetector(BaseProcess):
         )
 
         packet.broadcasts.append("circle_markers")
+
+        if self.display:
+            packet.display_hooks.append(self.display_hook)
 
         return packet